src/backend/pyproject.toml
[project]
name = "tero"
version = "0.1.0"
description = ""
authors = []
requires-python = ">=3.12,<3.14"
dependencies = [
    "fastapi (>=0.115.8,<0.116.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "openai (>=1.63.2,<2.0.0)",
    "sse-starlette (>=2.2.1,<3.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "psycopg[binary] (>=3.2.4,<4.0.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "greenlet (>=3.1.1,<4.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "aiohttp (>=3.11.12,<4.0.0)",
    "sqlmodel (>=0.0.22,<0.1.0)",
    "langchain-openai (>=0.3.6,<0.4.0)",
    "langchain-core (>=0.3.38,<0.4.0)",
    "langchain (>=0.3.19,<0.4.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "jsonschema (>=4.23.0,<5.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "langchain-postgres (>=0.0.13,<0.0.14)",
    "chardet (>=5.2.0,<6.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "langgraph (>=0.4.3,<0.5.0)",
    # Set langgraph-prebuilt to fixed version because newer versions have conflicts with current implementation
    "langgraph-prebuilt (>=0.1.8,<0.2.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "langchain-mcp-adapters (>=0.1.1,<0.2.0)",
    "cryptography (>=45.0.3,<46.0.0)",
    "azure-ai-documentintelligence (>=1.0.2,<2.0.0)",
    "tabulate (>=0.9.0,<0.10.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "pypdf (>=5.6.0,<6.0.0)",
    "langchain-community (>=0.3.24,<0.4.0)",
    "langchain-tavily (>=0.2.4,<0.3.0)",
    "langchain-google-community (>=2.0.7,<3.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)",
    "langchain-aws (>=0.2.27,<0.3.0)",
    "boto3 (>=1.39.4,<2.0.0)",
    "langchain-google-genai (>=2.1.8,<3.0.0)",
    "fastapi-mcp (>=0.4.0,<1.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "aiosmtplib (>=4.0.2,<5.0.0)",
    "xlrd (>=2.0.2,<3.0.0)",
    "openevals (>=0.1.0,<0.2.0)",
    "json-schema-to-pydantic (>=0.4.3,<0.5.0)",
    "transformers (>=4.57.1,<5.0.0)",
    "python-slugify (>=8.0.4,<9.0.0)",
    # Set mcp to fixed version because newer versions have conflicts with current implementation
    "mcp (>=1.19.0,<1.20.0)",
    "pypdfium2 (>=5.0.0,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
alembic = "^1.14.1"
pytest = "^8.3.4"
pytest-asyncio = "^0.26.0"
aiosqlite = "^0.21.0"
sqlparse = "^0.5.3"
freezegun = "^1.5.1"
testcontainers = {extras = ["postgres"], version = "^4.9.2"}
pyright = "^1.1.400"
alembic-postgresql-enum = "^1.7.0"
